% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{df.matched}
\alias{df.matched}
\title{compare the unique values of non-numerical columns of two data.frames}
\usage{
df.matched(df1, df2)
}
\arguments{
\item{df1, df2}{two data frames with identical column names and orders}
}
\value{
a matrix of following columns; NULL means matched.
\itemize{
\item column: column names of mis-match
\item df1-df2: additional unique values in df1
\item df2-df1: additional unique values in df1
}
}
\description{
Two data frames are matched if the unique values of each of their non-numerical columns are same.
Then a model trained from one data frame can predict the response in the other.
}
\examples{
# none
}
