% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{df.preprocess}
\alias{df.preprocess}
\title{preprocessing before training}
\usage{
df.preprocess(
  allDF = NULL,
  trainDF = NULL,
  testDF = NULL,
  prior.nzvCondX = c("n", "nx"),
  dummyEncoding = FALSE,
  prior.impute = c(NA, "missForest", "KNN"),
  classCol,
  posLabel = NULL,
  negLabel = NULL,
  oDir = NULL
)
}
\arguments{
\item{allDF}{NULL, data.frame; the data.frame of all the data}

\item{trainDF}{NULL, data.frame; the data.frame of the training data}

\item{testDF}{NULL, data.frame; the data.frame of the testing data}

\item{prior.nzvCondX}{NULL, char; remove NZV and conditionalX variables before training.
This can be used only if \code{caret::train} is not used in the model building.
Following are the available options:
* \code{NULL}: do not remove before training
* one of \verb{n, nx}:
+ n: NZV
+ nx: NZV and ConditionalX
if detected, removed variables are written to files \code{nzv.csv} and \code{conditionalX.txt}}

\item{dummyEncoding}{\code{cMethod \%in\% c('xgbTree','svmRadial')},logical; is dummy variable encoding needed before training;}

\item{prior.impute}{\code{c(NA,'missForest','KNN')}, char; impute missing values before training.
This can be used only if \code{caret::train} is not used in the model building.
Following are the available options:
* \code{NA}: do not impute before training
* one of \verb{missForest, KNN}: impute using \code{\link[missForest:missForest]{missForest::missForest()}} or \code{\link[DMwR2:knnImputation]{DMwR2::knnImputation()}}}

\item{classCol}{char; the column name of the class, i.e. response variable}

\item{posLabel, negLabel}{NULL, char; the positive and negative labels for two-class classification}

\item{oDir}{NULL, char; output directory}
}
\value{
a list containing \code{trainDF}, \code{testDF}, and \code{allDF} post processing.
}
\description{
options for preprocessing:
\itemize{
\item dummy variable encoding
\item modification to the whole dataset before training, including
\itemize{
\item imputation using \code{missForest} and \code{KNN}
\item detection and removal of \verb{Near Zero Variance (NZV)} and \code{ConditionalX} variables
Note that these modifications apply only in situations where \code{caret::train}
is not used in the model building; for example
\item ML method is \code{glm} where no cross validation (CV) is needed due to the absence of hyper-parameters
\item ML method is \code{randomForest} where the training is done using \code{\link[randomForest:tuneRF]{randomForest::tuneRF()}}
Otherwise, those modification should be specified in \code{caret::train(preProc)} to avoid
information leakage during training. Note that \code{missForest} is not available
in \code{preProc}.
}
}
}
\examples{
# none
}
