% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{draw.classic.MDS}
\alias{draw.classic.MDS}
\title{draw classic MDS plots}
\usage{
draw.classic.MDS(
  phenoDF,
  fMat = NULL,
  distMat = NULL,
  PC = 1:3,
  figMain = NULL,
  ofPre = NULL,
  colorCol = "black",
  doReturn = TRUE,
  ...
)
}
\arguments{
\item{phenoDF}{data.frame; row: subject; column: features to be used as color in the plot}

\item{fMat}{NULL, matrix; row: subjects at the same order as \code{phenoDF}; column: features}

\item{distMat}{distance matrix; subjects at the same order as \code{phenoDF};}

\item{PC}{1:3, integer vector; dimensions to plot}

\item{figMain}{NULL, char; title of the plots}

\item{ofPre}{NULL, char; prefix of output file name (including path)}

\item{colorCol}{'black', char; the color of each subject in the plot; can also be
a column in \code{phenoDF}}

\item{doReturn}{TRUE, logical; should a list of ggplot objects be returned?}

\item{...}{passed to \code{ggpubr::ggscatter}}
}
\value{
a list of ggplot objects if \code{doReturn = TRUE}
}
\description{
draw classic MDS plots
}
\examples{
# none
}
