% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{dummy.encode}
\alias{dummy.encode}
\title{encode categorical variables}
\usage{
dummy.encode(df1, useLevel = FALSE, ...)
}
\arguments{
\item{df1}{data frame; columns are variables.}

\item{useLevel}{FALSE, logical; if TRUE, for factor columns, the unique values
will be factor levels, instead of the actual levels present in the data.}

\item{...}{passed to \code{caret::dummyVars()}}
}
\value{
a new data.frame where all categorical columns are split into dummy
columns of values between -1 and 1.
}
\description{
convert categorical columns into numerical ones of full rank
}
\details{
It does the following:
\enumerate{
\item for nominal character variable of n categories, n-1 binary variables are created.
\item for nominal factor variable of n levels and m categories (m<=n),
\itemize{
\item if \code{useLevel = TRUE},  n-1 binary variables are created.
\item otherwise, m-1 binary variables are created.
}
\item for ordinal factor variable of n levels and m categories,
\itemize{
\item if m equals 2,  one variable assuming equal distance among levels are created
\item otherwise, n-1 variable assuming linear and 2 to n-1 degree of polynomial
relationship among levels are created.
}
\item cells of NA are kept.
}
}
\examples{
\dontrun{
k=matrix(1:10,nrow = 5)
colnames(k)=c('n1','n2')
k=as.data.frame(k)
k$c5=LETTERS[1:5]
k$c2na=c('a','a',NA,'b','b')
k$f2na=factor(k$c2na)
k$f2naL2O=factor(k$f2na,ordered = TRUE)
k$f2naL3=factor(k$c2na,levels = c('a','b','c'))
k$f2naL3O=factor(k$f2naL3,ordered = TRUE)
k$f3na=factor(c('a','a',NA,'b','c'))
k$f3naL3O=factor(k$f3na,ordered = TRUE)
k$f3naL4O=factor(k$f3na,levels = c('a','b','c','d'),ordered = TRUE)
k$f3naL5O=factor(k$f3na,levels = c('a','b','c','d','e'),ordered = TRUE)
k
dummy.encode(k)
}
}
