% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{list.concat}
\alias{list.concat}
\title{concatenate two lists}
\usage{
list.concat(list1, list2)
}
\arguments{
\item{list1}{the left list}

\item{list2}{the right list}
}
\value{
the final list
}
\description{
The final list contains items in both lists; but items in the left list that
share the same name as those in the right list are overwritten by the latter.
}
\examples{
# none
}
