% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{method2package}
\alias{method2package}
\title{get the name of R package for some frequently used methods}
\usage{
method2package(method)
}
\arguments{
\item{method}{char; the name of the method}
}
\value{
NA or the name of the package
}
\description{
preset method and package pairs
\itemize{
\item xgbTree = xgboost
\item randomForest = randomForest
\item rf = randomForest
\item parRF = randomForest
\item ranger = ranger
\item glmnet = glmnet
\item NaiveBayes = klaR
\item svmRadial = e1071
}
}
\examples{
# none
}
