% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{plotPredProbs}
\alias{plotPredProbs}
\title{plot prediction probability histogram}
\usage{
plotPredProbs(predProbs, trueLabels, posLabel)
}
\arguments{
\item{predProbs}{data.frame; a one column data.frame where the column name is
the positive label and rownames are subject IDs.}

\item{trueLabels}{char; the true class label}

\item{posLabel}{char; the positive label}
}
\value{
none
}
\description{
Produce back-to-back histogram of prediction probability distributions for
positive and negative labels
}
\examples{
#none
}
