% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{predStats.Categorical}
\alias{predStats.Categorical}
\title{compute the prediction stats of a model}
\usage{
predStats.Categorical(
  mFit,
  newData,
  classInd,
  posLabel = NULL,
  ofPre = NULL,
  predProbs = NULL,
  nLevels = NULL,
  ...
)
}
\arguments{
\item{mFit}{the model}

\item{newData}{data.frame; the new data}

\item{classInd}{integer; the index of the class column}

\item{posLabel}{NULL, char; the positive labels}

\item{ofPre}{NULL, char; the prefix of output file}

\item{predProbs}{NULL, data.frame; supplied prediction probability; column names
should be the levels in the response variable.}

\item{nLevels}{NULL, the number of levels in the class column.}

\item{...}{passed to \code{caret::predict()}}
}
\value{
a named numerical vector of stats
}
\description{
This function applies to the response variable of type factor. It computes the
prediction stats of a model and write the following files
}
\examples{
# none
}
