% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{predStats.numerical}
\alias{predStats.numerical}
\title{compute the prediction stats of a model}
\usage{
predStats.numerical(
  mFit,
  newData,
  classInd,
  ofPre = NULL,
  predVals = NULL,
  ...
)
}
\arguments{
\item{mFit}{the model}

\item{newData}{data.frame; the new data}

\item{classInd}{integer; the index of the class column}

\item{ofPre}{NULL, char; the prefix of output file}

\item{predVals}{NULL, data.frame; supplied prediction.}

\item{...}{passed to \code{caret::predict.train()}}

\item{posLabel}{NULL, char; the positive labels}
}
\value{
a named numerical vector of \verb{R2, RMSE, MAE}
}
\description{
This function computes the
prediction stats of a model on the target variable of type \code{numeric}.
}
\details{
Following files are the output:
\itemize{
\item cor.png
\item residule.png
\item prediction.csv
\item model.stats.txt
\item all.stats.rds
}
}
\examples{
# none
}
