% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{rm.NZV.condX}
\alias{rm.NZV.condX}
\title{remove low variance/varying and conditionalX features}
\usage{
rm.NZV.condX(df1, y = NULL, ofPre = NULL)
}
\arguments{
\item{df1}{data.frame; the input}

\item{y}{NULL, char; the column name of the response variable}

\item{ofPre}{NULL, char; the prefix of output file names:
* conditionalX.txt
* nzv.csv}
}
\value{
integer vector; the removed indices
}
\description{
low varying features are the non-numeric features where the most frequent value
accounts for >95\% of all values. low variance and conditionalX features are those defined in
\code{\link[caret:nearZeroVar]{caret::nearZeroVar()}} and \code{\link[caret:nearZeroVar]{caret::checkConditionalX()}}
}
\examples{
#none
}
