% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{writePredStats_bestTunes}
\alias{writePredStats_bestTunes}
\title{write the best-tune parameters to a file}
\usage{
writePredStats_bestTunes(mFit, fileName, statTable, aucVal = NA)
}
\arguments{
\item{mFit}{the final model}

\item{fileName}{char; output file name}

\item{statTable}{the statistic table}

\item{aucVal}{NA, numeric; prediction AUC}
}
\value{
none
}
\description{
write the best-tune parameters to a file
}
\examples{
# none
}
