% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML.funs.r
\name{predStats.Categorical}
\alias{predStats.Categorical}
\title{compute the prediction stats of a model}
\usage{
predStats.Categorical(
  mFit,
  newData,
  classInd,
  posLabel = NULL,
  ofPre = NULL,
  predProbs = NULL,
  nLevels = NULL,
  ...
)
}
\arguments{
\item{mFit}{the model}

\item{newData}{data.frame; the new data}

\item{classInd}{integer; the index of the class column}

\item{posLabel}{NULL, char; the positive labels}

\item{ofPre}{NULL, char; the prefix of output file}

\item{predProbs}{NULL, data.frame; supplied prediction probability; column names
should be the levels in the response variable.}

\item{nLevels}{NULL, the number of levels in the class column.}

\item{...}{passed to \code{caret::predict.train()}}
}
\value{
a named numerical vector of following stats:
\itemize{
\item AUC
\item accuracy and Kappa statistic values
\item the sensitivity, specificity, positive predictive value, negative
predictive value, precision, recall, F1, prevalence, detection rate,
detection prevalence and balanced accuracy for each class
}
}
\description{
This function computes the
prediction stats of a model on the target variable of type \code{factor}.
}
\details{
Following files are the output:
\itemize{
\item roc.pdf
\item prediction.hist.pdf
\item prediction.csv
\item confusion.table.txt
\item all.stats.rds
}
}
\examples{
# none
}
